chore(lld): Refactor device action structure
